+++ docs/CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Initialize repository 0604019 
     * repository created
 
-### Initialize repository 0105019 
-    * Project 2 added (LM35 to LCD/IC2)
+### Project 2 added 0105019 
+    *  LM35 to LCD/IC2
+
+### Project 3 Added 010619
+    * Tm1638 module (Key and LED) 8 LEDs and 8 Keys.